backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database (MySQL async)
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
alembic==1.14.0

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# HTTP client
httpx==0.28.0

# Excel / data ingestion
openpyxl==3.1.5
pandas==2.2.3

# Background tasks
celery==5.4.0
redis==5.2.1

# Push notifications (FCM)
firebase-admin==6.6.0

# Utilities
python-dotenv==1.0.1
Pillow==11.0.0
groq==0.11.0
pytz==2024.2

# Production Security & Logging
slowapi
structlog
apscheduler
cryptography
